--- 2-advanced/dubbo-samples-echo/src/main/java/org/apache/dubbo/samples/echo/EchoConsumer.java
@@ -1,38 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *   (the "License"); you may not use this file except in compliance with
- *   the License.  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-
-package org.apache.dubbo.samples.echo;
-
-import org.apache.dubbo.rpc.service.EchoService;
-import org.apache.dubbo.samples.echo.api.DemoService;
-
-import org.springframework.context.support.ClassPathXmlApplicationContext;
-
-public class EchoConsumer {
-
-    public static void main(String[] args) {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("spring/echo-consumer.xml");
-        context.start();
-        DemoService demoService = context.getBean("demoService", DemoService.class);
-
-        EchoService echoService = (EchoService) demoService;
-        String status = (String) echoService.$echo("OK");
-        System.out.println("echo result: " + status);
-    }
-}
